+++ discord/types/activity.py
@@ -41,9 +41,9 @@ class PartialPresenceUpdate(TypedDict):
 
 
 class ClientStatus(TypedDict, total=False):
-    desktop: str
-    mobile: str
-    web: str
+    desktop: StatusType
+    mobile: StatusType
+    web: StatusType
 
 
 class ActivityTimestamps(TypedDict, total=False):