+++ examples/guessing_game.py
@@ -1,3 +1,5 @@
+# This example requires the 'message_content' privileged intent to function.
+
 import discord
 import random
 import asyncio
@@ -30,5 +32,8 @@ class MyClient(discord.Client):
             else:
                 await message.channel.send(f'Oops. It is actually {answer}.')
 
-client = MyClient()
+intents = discord.Intents.default()
+intents.message_content = True
+
+client = MyClient(intents=intents)
 client.run('token')